# lab4/StudentService/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ModellingClub API
  version: 1.0.0
  description: >
    REST API for ModellingClub – a self-hosted web platform for scale-model
    and drone hobbyists.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BuildStatus:
      type: string
      enum:
        - draft
        - pending_review
        - published
        - rejected
        - archived
    BuildType:
      type: string
      enum:
        - fixed_wing
        - drone
        - helicopter
        - boat
        - car
        - scale_model
        - other
    BuildCreateRequest:
      type: object
      required:
        - title
        - type
      properties:
        title:
          type: string
          maxLength: 120
        type:
          $ref: '#/components/schemas/BuildType'
        description:
          type: string
        scale:
          type: string
          example: "1:72"
        wingspan_mm:
          type: integer
        frame_size_mm:
          type: integer
        num_motors:
          type: integer
        flight_controller:
          type: string
        max_takeoff_weight_g:
          type: integer
        rotor_diameter_mm:
          type: integer
        hull_length_mm:
          type: integer
        wheelbase_mm:
          type: integer
    Build:
      allOf:
        - $ref: '#/components/schemas/BuildCreateRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: Server-generated build ID.
            status:
              $ref: '#/components/schemas/BuildStatus'
            owner_id:
              type: string
              format: uuid
            score:
              type: number
              format: float
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    BuildList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Build'
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /builds:
    get:
      tags:
        - Builds
      summary: "UC-09: Browse and search published builds"
      parameters:
        - name: q
          in: query
          description: Full-text search (PostgreSQL tsvector)
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/BuildType'
        - name: author
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated list of published builds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildList'

    post:
      tags:
        - Builds
      summary: "UC-03: Create a new build (starts as draft)"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildCreateRequest'
      responses:
        '201':
          description: Build created
          headers:
            Location:
              description: Relative URL of the created build (contains the generated build ID).
              schema:
                type: string
                example: /builds/6f6d7154-7d2f-4a35-9e20-8536f6cbe6df
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Build'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...